+++ src/main/java/com/sahni/todoapi/ToDoAPI.java
@@ -1,4 +1,4 @@
-package com.sahni;
+package com.sahni.todoapi;
 
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
